# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REDPR документация АПИ",
        "description": "Документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/items": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "Поиск по каталогу",
                "description": "Поиск по каталогу",
                "operationId": "item",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "search": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Гипсокартон"
                                    },
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items/{id}": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "Отдельный товар",
                "description": "Отдельный товар",
                "operationId": "itemsolo",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Айди товара"
                                    },
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items/searchproducts": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "Отдельный товар",
                "description": "Отдельный товар",
                "operationId": "searchproducts",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "5"
                                    },
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/items/description": {
            "get": {
                "tags": [
                    "item"
                ],
                "summary": "Описание",
                "operationId": "description",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Категория",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Описание товара"
                    }
                }
            }
        },
        "/api/items/category/item": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "Поиск товаров по категории",
                "description": "Поиск товаров по категории",
                "operationId": "categoryitem",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "categoryID": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items/subcategory/item": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "Поиск товаров по подКатегории",
                "description": "Поиск товаров по подКатегории",
                "operationId": "subcategoryID",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "subcategoryID": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/popular": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "Популярные товары",
                "description": "Популярные товары",
                "operationId": "itempopular",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items/discount": {
            "post": {
                "tags": [
                    "item"
                ],
                "summary": "Популярные товары",
                "description": "Популярные товары",
                "operationId": "itemdiscount",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}